feat: Fix TopTag creates type to long

+++ src/main/java/com/innim/okkycopy/domain/board/dto/response/top_tag/TopTagResponse.java
@@ -15,7 +15,7 @@ import lombok.Setter;
 public class TopTagResponse {
 
     private String name;
-    private int cnt;
+    private long cnt;
 
     public static TopTagResponse from(TopTag topTag) {
         return TopTagResponse.builder()
